requirements.txt
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
google-api-core==1.31.0
google-auth==1.33.1
googleapis-common-protos==1.53.0
idna==3.2
numpy==1.21.1
packaging==21.0
pandas==1.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
urllib3==1.26.6
